requirements.txt
feedparser>=6.0.11,<7
python-dotenv>=1.0.1,<2
requests>=2.32.0,<3
